rename batch_manager to batch
